refactor(project): move template filter arity
