+++ crates/djls-project/src/lib.rs
@@ -56,8 +56,8 @@ pub use specs::ExtractedDiagnosticConstraint;
 pub use specs::ExtractedDiagnosticMessage;
 pub use specs::ExtractedMessageArg;
 pub use specs::ExtractedMessageTemplate;
-pub use specs::FilterArity;
-pub use specs::FilterArityMap;
+pub use templates::FilterArity;
+pub use templates::FilterArityMap;
 pub use specs::KnownOptions;
 pub use specs::ModelGraph;
 pub use specs::RequiredKeyword;
