+++ crates/djls-project/src/templates/filters.rs
@@ -1,6 +1,23 @@
 use ruff_python_ast::StmtFunctionDef;
+use rustc_hash::FxHashMap;
+use serde::Deserialize;
+use serde::Serialize;
 
-use crate::specs::types::FilterArity;
+use crate::templates::SymbolKey;
+
+pub type FilterArityMap = FxHashMap<SymbolKey, FilterArity>;
+
+/// Filter argument arity extracted from the filter function's signature.
+///
+/// Django filters receive the value being filtered as their first argument.
+/// Some filters accept an additional argument (e.g., `{{ value|default:"nothing" }}`).
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct FilterArity {
+    /// Whether the filter expects an argument after the colon.
+    pub expects_arg: bool,
+    /// Whether the argument is optional (has a default value).
+    pub arg_optional: bool,
+}
 
 /// Extract filter argument arity from a filter function's signature.
 ///
@@ -68,6 +85,35 @@ mod tests {
     use super::*;
     use crate::specs::testing::django_function;
     use crate::specs::testing::find_function_in_source;
+
+    #[test]
+    fn filter_arity_no_arg() {
+        let arity = FilterArity {
+            expects_arg: false,
+            arg_optional: false,
+        };
+        assert!(!arity.expects_arg);
+    }
+
+    #[test]
+    fn filter_arity_required_arg() {
+        let arity = FilterArity {
+            expects_arg: true,
+            arg_optional: false,
+        };
+        assert!(arity.expects_arg);
+        assert!(!arity.arg_optional);
+    }
+
+    #[test]
+    fn filter_arity_optional_arg() {
+        let arity = FilterArity {
+            expects_arg: true,
+            arg_optional: true,
+        };
+        assert!(arity.expects_arg);
+        assert!(arity.arg_optional);
+    }
 
     // No-arg filters (value only)
